+++ amdsmi_cli/amdsmi_commands.py
@@ -98,12 +98,22 @@ class AMDSMICommands():
             sys.exit(-1)
 
 
-    def version(self, args):
+    def version(self, args, gpu_version=None):
         """Print Version String
 
         Args:
             args (Namespace): Namespace containing the parsed CLI args
         """
+
+        if gpu_version:
+            args.gpu_version = gpu_version
+        # if cpu_version:
+        #     args.cpu_version = cpu_version
+        # # if no args are given, display everything
+        # if not gpu_version and not cpu_version:
+        #     args.gpu_version = True
+        #     args.cpu_version = True
+
         try:
             amdsmi_lib_version = amdsmi_interface.amdsmi_get_lib_version()
             amdsmi_lib_version_str = f"{amdsmi_lib_version['year']}.{amdsmi_lib_version['major']}.{amdsmi_lib_version['minor']}.{amdsmi_lib_version['release']}"
@@ -116,10 +126,38 @@ class AMDSMICommands():
         self.logger.output['amdsmi_library_version'] = f'{amdsmi_lib_version_str}'
         self.logger.output['rocm_version'] = f'{rocm_version_str}'
 
+        if args.gpu_version:
+            try:
+                gpus = amdsmi_interface.amdsmi_get_processor_handles()
+                if gpus:
+                    gpu_version_info = amdsmi_interface.amdsmi_get_gpu_driver_info(gpus[0])
+                    gpu_version_str = gpu_version_info['driver_version']
+                else:
+                    gpu_version_str = "N/A"
+            except amdsmi_exception.AmdSmiLibraryException as e:
+                gpu_version_str = e.get_error_info()
+            self.logger.output['amdgpu_version'] = gpu_version_str
+
+        # if args.cpu_version:
+        #     try:
+        #         cpus = amdsmi_interface.amdsmi_get_cpusocket_handles()
+        #         if cpus:
+        #             cpu_version_info = amdsmi_interface.amdsmi_get_amd_hsmp_deriver_version(cpus[0])
+        #             cpu_version_str = cpu_version_info['driver_version']
+        #         else:
+        #             cpu_version_str = "N/A"
+        #     except amdsmi_exception.AmdSmiLibraryException as e:
+        #         cpu_version_str = e.get_error_info()
+        #     self.logger.output['amd_hsmp_driver_version'] = cpu_version_str
+
         if self.logger.is_human_readable_format():
             human_readable_output = f"AMDSMI Tool: {__version__} | " \
                                     f"AMDSMI Library version: {amdsmi_lib_version_str} | " \
                                     f"ROCm version: {rocm_version_str}"
+            if args.gpu_version:
+                human_readable_output = human_readable_output + f" | amdgpu version: {gpu_version_str}"
+            # if args.cpu_version:
+            #     human_readable_output = human_readable_output + f" | amd_hsmp driver version: {cpu_version_str}"
             # Custom human readable handling for version
             if self.logger.destination == 'stdout':
                 print(human_readable_output)